+++ extensions/sportssync/src/basketball-scores-and-schedule.tsx
@@ -1,6 +1,7 @@
 import { Detail, List, Color, Action, ActionPanel } from "@raycast/api";
 import { useFetch } from "@raycast/utils";
 import { useState } from "react";
+import getPastAndFutureDays from "./utils/getDateRange";
 
 interface Competitor {
   team: {
@@ -33,6 +34,11 @@ interface Game {
   links: { href: string }[];
 }
 
+interface DayItems {
+  title: string;
+  games: JSX.Element[];
+}
+
 interface Response {
   events: Game[];
   day: { date: string };
@@ -41,13 +47,31 @@ interface Response {
 export default function scoresAndSchedule() {
   // Fetch NBA Stats
 
+  const dateRange = getPastAndFutureDays(new Date());
+
   const [currentLeague, displaySelectLeague] = useState("NBA Games");
   const { isLoading: nbaScheduleStats, data: nbaScoresAndSchedule } = useFetch<Response>(
-    "https://site.api.espn.com/apis/site/v2/sports/basketball/nba/scoreboard",
+    `https://site.api.espn.com/apis/site/v2/sports/basketball/nba/scoreboard?dates=${dateRange}`,
   );
 
+  const nbaDayItems: DayItems[] = [];
   const nbaGames = nbaScoresAndSchedule?.events || [];
-  const nbaItems = nbaGames.map((nbaGame, index) => {
+
+  nbaGames.forEach((nbaGame, index) => {
+    const gameDate = new Date(nbaGame.date);
+    const nbaGameDay = gameDate.toLocaleDateString([], {
+      dateStyle: "medium",
+    });
+
+    if (!nbaDayItems.find((nbaDay) => nbaDay.title === nbaGameDay)) {
+      nbaDayItems.push({
+        title: nbaGameDay,
+        games: [],
+      });
+    }
+
+    const nbaDay = nbaDayItems.find((nbaDay) => nbaDay.title === nbaGameDay);
+
     const gameTime = new Date(nbaGame.date).toLocaleTimeString([], {
       hour: "2-digit",
       minute: "2-digit",
@@ -74,7 +98,7 @@ export default function scoresAndSchedule() {
       accessoryColor = Color.Orange;
     }
 
-    return (
+    nbaDay?.games.push(
       <List.Item
         key={index}
         title={`${nbaGame.name}`}
@@ -98,17 +122,34 @@ export default function scoresAndSchedule() {
             )}
           </ActionPanel>
         }
-      />
+      />,
     );
   });
 
   // Fetch WNBA Stats
 
   const { isLoading: wnbaScheduleStats, data: wnbaScoresAndSchedule } = useFetch<Response>(
-    "https://site.api.espn.com/apis/site/v2/sports/basketball/wnba/scoreboard",
+    `https://site.api.espn.com/apis/site/v2/sports/basketball/wnba/scoreboard?dates=${dateRange}`,
   );
+
+  const wnbaDayItems: DayItems[] = [];
   const wnbaGames = wnbaScoresAndSchedule?.events || [];
-  const wnbaItems = wnbaGames.map((wnbaGame, index) => {
+
+  wnbaGames.forEach((wnbaGame, index) => {
+    const gameDate = new Date(wnbaGame.date);
+    const wnbaGameDay = gameDate.toLocaleDateString([], {
+      dateStyle: "medium",
+    });
+
+    if (!wnbaDayItems.find((wnbaDay) => wnbaDay.title === wnbaGameDay)) {
+      wnbaDayItems.push({
+        title: wnbaGameDay,
+        games: [],
+      });
+    }
+
+    const wnbaDay = wnbaDayItems.find((wnbaDay) => wnbaDay.title === wnbaGameDay);
+
     const gameTime = new Date(wnbaGame.date).toLocaleTimeString([], {
       hour: "2-digit",
       minute: "2-digit",
@@ -135,7 +176,7 @@ export default function scoresAndSchedule() {
       accessoryColor = Color.Orange;
     }
 
-    return (
+    wnbaDay?.games.push(
       <List.Item
         key={index}
         title={`${wnbaGame.name}`}
@@ -159,7 +200,7 @@ export default function scoresAndSchedule() {
             )}
           </ActionPanel>
         }
-      />
+      />,
     );
   });
 
@@ -167,8 +208,17 @@ export default function scoresAndSchedule() {
     return <Detail isLoading={true} />;
   }
 
-  const nbaGamesDate = nbaScoresAndSchedule?.day?.date ?? "No date available";
-  const wnbaGamesDate = wnbaScoresAndSchedule?.day?.date ?? "No date available";
+  nbaDayItems.sort((a, b) => {
+    const dateA = new Date(a.title);
+    const dateB = new Date(b.title);
+    return dateA.getTime() - dateB.getTime();
+  });
+
+  wnbaDayItems.sort((a, b) => {
+    const dateA = new Date(a.title);
+    const dateB = new Date(b.title);
+    return dateA.getTime() - dateB.getTime();
+  });
 
   return (
     <List
@@ -183,23 +233,29 @@ export default function scoresAndSchedule() {
     >
       {currentLeague === "NBA" && (
         <>
-          <List.Section
-            title={`${nbaGamesDate}`}
-            subtitle={`${nbaItems.length} Game${nbaItems.length !== 1 ? "s" : ""}`}
-          >
-            {nbaItems}
-          </List.Section>
+          {nbaDayItems.map((nbaDay, index) => (
+            <List.Section
+              key={index}
+              title={`${nbaDay.title}`}
+              subtitle={`${nbaDay.games.length} Game${nbaDay.games.length !== 1 ? "s" : ""}`}
+            >
+              {nbaDay.games}
+            </List.Section>
+          ))}
         </>
       )}
 
       {currentLeague === "WNBA" && (
         <>
-          <List.Section
-            title={`${wnbaGamesDate}`}
-            subtitle={`${wnbaItems.length} Game${wnbaItems.length !== 1 ? "s" : ""}`}
-          >
-            {wnbaItems}
-          </List.Section>
+          {wnbaDayItems.map((wnbaDay, index) => (
+            <List.Section
+              key={index}
+              title={`${wnbaDay.title}`}
+              subtitle={`${wnbaDay.games.length} Game${wnbaDay.games.length !== 1 ? "s" : ""}`}
+            >
+              {wnbaDay.games}
+            </List.Section>
+          ))}
         </>
       )}
     </List>