+++ extensions/sportssync/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Sportssync Changelog
 
+<<<<<<< HEAD
+## [Implemented past and future scores] - {PR_MERGE_DATE}
+
+- Updated the NHL Standings command to use divisions instead of conferences
+- Updated each scores and schedule command to now show the scores for the past 3 days, and the games for the next 5 days
+
+## [Added a bunch of new feedback] - 2025-02-13
+=======
 ## [Added a bunch of new feedback] - 2025-02-14
+>>>>>>> contributions/merge-1739918385494
 
 - Added a new league: Champions League
 - Fixed an issue causing the record in nhl standings command to show up as undefined
@@ -9,9 +18,7 @@
 - Updated the F1 Scores and Schedule command title: F1 Results and Schedule
 - Added proper support for 4 Nations for the NHL Scores and Schedule Command
 
-## [Initial Version] - 2025-02-13
-
-Initial Sportssync extension along the following features:
+## [Initial Version] - 2025-02-12
 
 - View upcoming games and live scores
 - View up to date standings with in depth team stats